Make pricing the index page.
